refactor: move all fix-pr functionality to 1 file
